requirements.txt
numpy==1.21.5
opencv_python_headless==4.8.0.76
Pillow==10.0.0
scikit_learn==1.3.0
tensorflow==2.13.0
